# internal/transport/apiserver/gen/api.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package gen

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ObjError Standard error
type ObjError struct {
	Code      string `json:"code"`
	Message   string `json:"message"`
	RequestId string `json:"request_id"`
}

// ObjUser Representation of user
type ObjUser struct {
	FirstName *string `json:"first_name,omitempty"`
	LastName  *string `json:"last_name,omitempty"`
	TgChatId  *int64  `json:"tg_chat_id,omitempty"`
	TgId      *int64  `json:"tg_id,omitempty"`
	Username  *string `json:"username,omitempty"`
}

// UserId defines model for UserId.
type UserId = openapi_types.UUID

// RespCreateUser defines model for RespCreateUser.
type RespCreateUser struct {
	Data struct {
		// UserId Unique user identifier (UUID)
		UserId openapi_types.UUID `json:"user_id"`
	} `json:"data"`
}

// RespError defines model for RespError.
type RespError struct {
	// Error Standard error
	Error ObjError               `json:"error"`
	Meta  map[string]interface{} `json:"meta"`
}

// RespGetUser defines model for RespGetUser.
type RespGetUser struct {
	// Data Representation of user
	Data ObjUser `json:"data"`
}

// RespUpdateUser defines model for RespUpdateUser.
type RespUpdateUser struct {
	// Data Representation of user
	Data ObjUser `json:"data"`
}

// ReqCreateUser defines model for ReqCreateUser.
type ReqCreateUser struct {
	FirstName *string `json:"first_name,omitempty"`
	LastName  *string `json:"last_name,omitempty"`
	TgChatId  int64   `json:"tg_chat_id"`
	TgId      int64   `json:"tg_id"`
	Username  *string `json:"username,omitempty"`
}

// ReqUpdateUser defines model for ReqUpdateUser.
type ReqUpdateUser struct {
	FirstName *string `json:"first_name,omitempty"`
	LastName  *string `json:"last_name,omitempty"`
	TgChatId  *int64  `json:"tg_chat_id,omitempty"`
	TgId      *int64  `json:"tg_id,omitempty"`
	Username  *string `json:"username,omitempty"`
}

// V1CreateUserJSONBody defines parameters for V1CreateUser.
type V1CreateUserJSONBody struct {
	FirstName *string `json:"first_name,omitempty"`
	LastName  *string `json:"last_name,omitempty"`
	TgChatId  int64   `json:"tg_chat_id"`
	TgId      int64   `json:"tg_id"`
	Username  *string `json:"username,omitempty"`
}

// V1CreateUserJSONRequestBody defines body for V1CreateUser for application/json ContentType.
type V1CreateUserJSONRequestBody V1CreateUserJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Create new user
	// (POST /api/v1/users)
	V1CreateUser(w http.ResponseWriter, r *http.Request)
	// Fetch user by id
	// (GET /api/v1/users/{id})
	V1GetUser(w http.ResponseWriter, r *http.Request, id UserId)
	// Update user info
	// (PUT /api/v1/users/{id})
	V1UpdateUser(w http.ResponseWriter, r *http.Request, id UserId)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Create new user
// (POST /api/v1/users)
func (_ Unimplemented) V1CreateUser(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Fetch user by id
// (GET /api/v1/users/{id})
func (_ Unimplemented) V1GetUser(w http.ResponseWriter, r *http.Request, id UserId) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Update user info
// (PUT /api/v1/users/{id})
func (_ Unimplemented) V1UpdateUser(w http.ResponseWriter, r *http.Request, id UserId) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// V1CreateUser operation middleware
func (siw *ServerInterfaceWrapper) V1CreateUser(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.V1CreateUser(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// V1GetUser operation middleware
func (siw *ServerInterfaceWrapper) V1GetUser(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id UserId

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.V1GetUser(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// V1UpdateUser operation middleware
func (siw *ServerInterfaceWrapper) V1UpdateUser(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id UserId

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.V1UpdateUser(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1/users", wrapper.V1CreateUser)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/users/{id}", wrapper.V1GetUser)
	})
	r.Group(func(r chi.Router) {
		r.Put(options.BaseURL+"/api/v1/users/{id}", wrapper.V1UpdateUser)
	})

	return r
}

type RespCreateUserJSONResponse struct {
	Data struct {
		// UserId Unique user identifier (UUID)
		UserId openapi_types.UUID `json:"user_id"`
	} `json:"data"`
}

type RespErrorJSONResponse struct {
	// Error Standard error
	Error ObjError               `json:"error"`
	Meta  map[string]interface{} `json:"meta"`
}

type RespGetUserJSONResponse struct {
	// Data Representation of user
	Data ObjUser `json:"data"`
}

type RespUpdateUserJSONResponse struct {
	// Data Representation of user
	Data ObjUser `json:"data"`
}

type V1CreateUserRequestObject struct {
	Body *V1CreateUserJSONRequestBody
}

type V1CreateUserResponseObject interface {
	VisitV1CreateUserResponse(w http.ResponseWriter) error
}

type V1CreateUser200JSONResponse struct{ RespCreateUserJSONResponse }

func (response V1CreateUser200JSONResponse) VisitV1CreateUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type V1CreateUserdefaultJSONResponse struct {
	Body struct {
		// Error Standard error
		Error ObjError               `json:"error"`
		Meta  map[string]interface{} `json:"meta"`
	}
	StatusCode int
}

func (response V1CreateUserdefaultJSONResponse) VisitV1CreateUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)

	return json.NewEncoder(w).Encode(response.Body)
}

type V1GetUserRequestObject struct {
	Id UserId `json:"id"`
}

type V1GetUserResponseObject interface {
	VisitV1GetUserResponse(w http.ResponseWriter) error
}

type V1GetUser200JSONResponse struct{ RespGetUserJSONResponse }

func (response V1GetUser200JSONResponse) VisitV1GetUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type V1GetUserdefaultJSONResponse struct {
	Body struct {
		// Error Standard error
		Error ObjError               `json:"error"`
		Meta  map[string]interface{} `json:"meta"`
	}
	StatusCode int
}

func (response V1GetUserdefaultJSONResponse) VisitV1GetUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)

	return json.NewEncoder(w).Encode(response.Body)
}

type V1UpdateUserRequestObject struct {
	Id UserId `json:"id"`
}

type V1UpdateUserResponseObject interface {
	VisitV1UpdateUserResponse(w http.ResponseWriter) error
}

type V1UpdateUser200JSONResponse struct{ RespUpdateUserJSONResponse }

func (response V1UpdateUser200JSONResponse) VisitV1UpdateUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type V1UpdateUserdefaultJSONResponse struct {
	Body struct {
		// Error Standard error
		Error ObjError               `json:"error"`
		Meta  map[string]interface{} `json:"meta"`
	}
	StatusCode int
}

func (response V1UpdateUserdefaultJSONResponse) VisitV1UpdateUserResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(response.StatusCode)

	return json.NewEncoder(w).Encode(response.Body)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Create new user
	// (POST /api/v1/users)
	V1CreateUser(ctx context.Context, request V1CreateUserRequestObject) (V1CreateUserResponseObject, error)
	// Fetch user by id
	// (GET /api/v1/users/{id})
	V1GetUser(ctx context.Context, request V1GetUserRequestObject) (V1GetUserResponseObject, error)
	// Update user info
	// (PUT /api/v1/users/{id})
	V1UpdateUser(ctx context.Context, request V1UpdateUserRequestObject) (V1UpdateUserResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// V1CreateUser operation middleware
func (sh *strictHandler) V1CreateUser(w http.ResponseWriter, r *http.Request) {
	var request V1CreateUserRequestObject

	var body V1CreateUserJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.V1CreateUser(ctx, request.(V1CreateUserRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "V1CreateUser")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(V1CreateUserResponseObject); ok {
		if err := validResponse.VisitV1CreateUserResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// V1GetUser operation middleware
func (sh *strictHandler) V1GetUser(w http.ResponseWriter, r *http.Request, id UserId) {
	var request V1GetUserRequestObject

	request.Id = id

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.V1GetUser(ctx, request.(V1GetUserRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "V1GetUser")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(V1GetUserResponseObject); ok {
		if err := validResponse.VisitV1GetUserResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// V1UpdateUser operation middleware
func (sh *strictHandler) V1UpdateUser(w http.ResponseWriter, r *http.Request, id UserId) {
	var request V1UpdateUserRequestObject

	request.Id = id

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.V1UpdateUser(ctx, request.(V1UpdateUserRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "V1UpdateUser")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(V1UpdateUserResponseObject); ok {
		if err := validResponse.VisitV1UpdateUserResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}